One additional criteria for dynamic breaks and speed-up of Hellinger
Addition of bigg breaks that comes from the analysis of the cumulative distribution of the joinx matrix.
They are simply joined together with the usual dynamical one.
As opposed to these, they should be able to identify entropic basins-states at the very right of the basins (e.g. Nbutane no LeavesFolding).
The previous method to compute Hellinger has been replaced by a faster homemade one. The bottleneck removed basically was DiscreteDistribution().
Eventually, removal of big data.frame not used after a certain point in the function.